build: Set/unset dead link checker in Git hooks
It's usually tedious and time-consuming to run the dead link checker
on every commit, particularly if one is working on a local topic
branch that gets frequent commits. So, update the pre-commit hook
script to disable it (unless it has been specifically enabled in the
environment already). However, enable it unconditionally in the
pre-push hook so that we don't create broken pull requests.